Add int cast for File::createThumb
createThumb needs a int
There are some more int cast on multiply with 1.5 in this code

+++ includes/TimelessTemplate.php
@@ -1023,7 +1023,7 @@ class TimelessTemplate extends BaseTemplate {
 
 				if ( $svg ) {
 					$logoData['1x'] = $file->createThumb( $logoData['width'] );
-					$logoData['1.5x'] = $file->createThumb( $logoData['width'] * 1.5 );
+					$logoData['1.5x'] = $file->createThumb( (int)( $logoData['width'] * 1.5 ) );
 					$logoData['2x'] = $file->createThumb( $logoData['width'] * $large );
 				} elseif ( $file->mustRender() ) {
 					$logoData['1x'] = $file->createThumb( $logoData['width'] );
